refactor: dao 제거
