+++ src/test/java/woowacourse/shoppingcart/dao/CustomerDaoTest.java
@@ -1,13 +1,17 @@
 package woowacourse.shoppingcart.dao;
 
+import javax.sql.DataSource;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase.Replace;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.TestConstructor;
 import org.springframework.test.context.jdbc.Sql;
+import woowacourse.auth.domain.Customer;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -19,8 +23,16 @@ public class CustomerDaoTest {
 
     private final CustomerDao customerDao;
 
-    public CustomerDaoTest(JdbcTemplate jdbcTemplate) {
-        customerDao = new CustomerDao(jdbcTemplate);
+    public CustomerDaoTest(DataSource dataSource) {
+        customerDao = new CustomerDao(dataSource);
+    }
+
+    @Test
+    @DisplayName("회원을 저장한다.")
+    void save() {
+        Customer customer = customerDao.save(new Customer("레넌", "rennon@woowa.com", "1234"));
+
+        assertThat(customer.getId()).isEqualTo(1L);
     }
 
     @DisplayName("username을 통해 아이디를 찾으면, id를 반환한다.")
